+++ ProyectoFinalJava/src/Models/Mecanico.java
@@ -10,7 +10,7 @@ public class Mecanico extends Persona implements IJson {
     private String contrasenia;
 
 
-    public Mecanico(String nombre, String apellido, int dni, int telefono, String email, String usuario, String contrasenia) {
+    public Mecanico(String nombre, String apellido, long dni, long telefono, String email, String usuario, String contrasenia) {
         super(nombre, apellido, dni, telefono, email);
         this.usuario = usuario;
         this.contrasenia = Seguridad.hashearContrasenia(contrasenia);
@@ -95,8 +95,8 @@ public class Mecanico extends Persona implements IJson {
             ///  Los atributos que tienen como herencia de Persona
             mecanico.setNombre(object.getString("nombre"));
             mecanico.setApellido(object.getString("apellido"));
-            mecanico.setDni(object.getInt("dni"));
-            mecanico.setTelefono(object.getInt("telefono"));
+            mecanico.setDni(object.getLong("dni"));
+            mecanico.setTelefono(object.getLong("telefono"));
             mecanico.setEmail(object.getString("email"));
 
             // Campos propios de Mecanico
